move devices to folder
