+++ devices/baseDevice.js
